% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{execute_cibersort}
\alias{execute_cibersort}
\title{Execute cibersort using Rserve}
\usage{
execute_cibersort(input_file, output_file, absolute_mode = FALSE,
  tmpdir = tempdir(), cibersort_home = "~/cibersort",
  training_input = file.path(cibersort_home, "LM22.txt"),
  rserve_log = file.path(tmpdir, "Rserve_log.txt"),
  error_file = file.path(tmpdir, "cibersort_errors.txt"))
}
\arguments{
\item{input_file}{path to input file on disk}

\item{output_file}{path where output should be stored (on disk)}

\item{tmpdir}{directory to use for intermediate results}

\item{cibersort_home}{directory into which CIBERSORT_package.zip has been unzipped}

\item{training_input}{path to gene signatures for cell types (defaults to LM22)}

\item{rserve_log}{path in which to store Rserve stderr (contents printed on error)}

\item{error_file}{path in which to store cibersort stderr (contents printed on error)}
}
\value{
cibersort results in a data.frame
}
\description{
Execute cibersort using Rserve
}
\note{
used by \code\{run_cibersort\} to call out to cibersort.
}
\seealso{
\code\{run_cibersort\} which takes an expdata object rather than a filepath
}
