% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{expdata_as_df}
\alias{expdata_as_df}
\title{convert expdaat (as GxS matrix) to tbl_df format}
\usage{
expdata_as_df(expmat, sampleinfo)
}
\arguments{
\item{expmat}{(matrix) expression data in GxS format}

\item{sampleinfo}{(tbL_df or data.frame) sampleinfo with `_SAMPLE_ID` column corresponding to column names in expmat}
}
\value{
expdata (tbl_df) containing expression data, with sampleinfo attribute
}
\description{
convert expdaat (as GxS matrix) to tbl_df format
}
