% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{expdata_as_matrix}
\alias{expdata_as_matrix}
\title{convert expdata (as GxS tbl_df) to matrix format}
\usage{
expdata_as_matrix(df, trans = NULL)
}
\arguments{
\item{df}{(tbl_df or data.frame) a GxS data frame containing expression data}

\item{trans}{(function) optional transformation to apply to values in matrix}
}
\value{
matrix of expression data, posssibly transformed values
}
\description{
convert expdata (as GxS tbl_df) to matrix format
}
