% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{filter_batch}
\alias{filter_batch}
\title{filter expression data and sample info to desired batches}
\usage{
filter_batch(df, batch_names)
}
\arguments{
\item{df}{(tbl_df) SxG df with expression data with columns indicating sample IDs, where sampleinfo stored in sampleinfo attr}

\item{batch_names}{(character vector) names of batches to keep}
}
\value{
SxG df with expression data and sampleinfo in sampleinfo attr
}
\description{
filter expression data and sample info to desired batches
}
