% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{filter_expdata}
\alias{filter_expdata}
\title{helper function to filter gxs expdata more efficiently
than can be done using data.frame tools}
\usage{
filter_expdata(df, fun = function(x) {     max(x) > 0 }, trans = NULL)
}
\arguments{
\item{df}{(tbl_df or data.frame) GxS expression data with a column `_GENE`}

\item{fun}{(function) optional function to be applied to each row (GENE) of expression data
returns a boolean value (TRUE: keep, FALSE: discard)}

\item{trans}{(function) optional transformation to apply to expression data AFTER filtering, for convenience}
}
\value{
filtered df (tbl_df) containing filtered and/or transformed GxS expression data
}
\description{
helper function to filter gxs expdata more efficiently
than can be done using data.frame tools
}
