% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{filter_genes}
\alias{filter_genes}
\title{Filter GxS expression data to specific list of genes
If `include_random` is true, will optionally select random genes
to round out the comparison}
\usage{
filter_genes(df, n_genes = 900,
  genelist = readr::read_tsv("/data/rcc-immuneinf/marker_genes.txt", col_names
  = c("_GENE")), include_random = TRUE)
}
\arguments{
\item{df}{(tbl_df) SxG df with expression data, where sampleinfo stored in sampleinfo attr}

\item{n_genes}{(int) desired total number of genes (irrelevant if include_random == FALSE)}

\item{genelist}{(tbl_df) contains column `_GENE` containing list of genes to filter to}

\item{include_random}{(bool) whether to sample from remaining genes to reach `n_genes`}
}
\value{
tbl_df containing filtered expression data, with sampleinfo attr
}
\description{
Filter GxS expression data to specific list of genes
If `include_random` is true, will optionally select random genes
to round out the comparison
}
