% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{load_all_expdata}
\alias{load_all_expdata}
\title{load several expdata files all at once}
\usage{
load_all_expdata(path, batch = NULL, ...)
}
\arguments{
\item{path}{(vector or list) filepaths containing expression matrices}

\item{batch}{(vector or list) batch name corresponding to each filepath}

\item{...}{other params to \code{load_expdata}, each of which can be a vector or singleton}
}
\value{
list of imported dfs returned by load_expdata
}
\description{
load several expdata files all at once
}
\note{
this function is a lightweight wrapper around \code{load_expdata}.
      Please see ?merge_expdata for more info on parameters
}
