% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{load_expdata}
\alias{load_expdata}
\title{loads expression data in GxS format, with one row per gene & one column per sample
returns a tbl_df also in GxS format, with sample-batch-assignment in attr `sampleinfo`}
\usage{
load_expdata(path, gene_col = "Gene_symbol", data_loader = readr::read_tsv,
  batch = 1, filter = NULL, trans = NULL, ...)
}
\arguments{
\item{path}{(chr) path to expression matrix}

\item{gene_col}{(chr) name of column containing gene symbol}

\item{data_loader}{(function) function to use when loading data}

\item{batch}{(chr) optional description to give this batch}

\item{...}{extra args to data_loader}

\item{expression_filter}{(dbl) optional value by which to filter expression (if NULL, no filter)}
}
\value{
tbl_df containing expression data, with columns `_SAMPLE_ID` and `_BATCH`
}
\description{
loads expression data in GxS format, with one row per gene & one column per sample
returns a tbl_df also in GxS format, with sample-batch-assignment in attr `sampleinfo`
}
\note{
Expects data in format prepared by prep_gene_expression_matrix.R

Genes are filtered to remove those with 0 expression across all samples
      these can be filled in with 0 if expressed in at least one batch on combination
}
