% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{load_lm22_genes}
\alias{load_lm22_genes}
\title{Helper function to load lm22 genes from cibersort_home
and prepare in format for `filter_genes`}
\usage{
load_lm22_genes(cibersort_home = "~/cibersort")
}
\arguments{
\item{cibersort_home}{(str) path to cibersort home dir}
}
\description{
Helper function to load lm22 genes from cibersort_home
and prepare in format for `filter_genes`
}
