% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{merge_expdata}
\alias{merge_expdata}
\title{merge expdata dfs together, combining sampleinfo}
\usage{
merge_expdata(dflist, fill = NULL, filter = function(x) {     max(x) > 0 })
}
\arguments{
\item{dflist}{(list) list of dataframes imported using load_expdata}

\item{fill}{(value or NULL) how to fill in NA values, if at all}

\item{filter}{(function) function to apply to rows, filtering genes/transcripts from result}
}
\description{
merge expdata dfs together, combining sampleinfo
}
