% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_expression_matrix.R
\name{prep_expression_matrix}
\alias{prep_expression_matrix}
\title{Prepare expression matrix of transcript- or gene-specific data
given a directory containing kallisto output (one folder per sample)}
\usage{
prep_expression_matrix(root, output_arg, skip_missing = TRUE,
  cohort = basename(root), metric = "no", abundanceCol = "TPM",
  by = "Gene_symbol", by_gene = TRUE, suffix = "tsv",
  filename = glue::glue("abundance.{suffix}"),
  output_writer = readr::write_tsv)
}
\arguments{
\item{root}{(chr) filepath to 'root' directory containing kallisto output}

\item{output_arg}{(chr) argument to `output_writer`, typically the filename to be written}

\item{skip_missing}{(bool) whether to report & ignore directories not containing output}

\item{metric}{(chr) which metric to capture from kallisto. One of 'scaledTPM', 'no', (see `?tximport` for details)}

\item{by}{(chr) description / name for column in output matrix containing gene or transcript}

\item{by_gene}{(bool) whether to organize results by gene (TRUE) or transcript (FALSE)}

\item{suffix}{(chr) file suffix to read in (either `tsv` or `h5` -- h5 gives errors on some platforms)}

\item{filename}{(chr) name of kallisto output file to ingest}

\item{output_writer}{(function) function to use when saving results.}
}
\description{
Prepare expression matrix of transcript- or gene-specific data
given a directory containing kallisto output (one folder per sample)
}
