% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcctils_expression-data.R
\docType{data}
\name{rcctils_expression}
\alias{rcctils_expression}
\title{Gene expression matrices for TILs isolated from RCC patients (E-MTAB-5640)}
\format{an object of class \code{"tbl_df"}}
\usage{
data(genelist)
}
\description{
Used as a training dataset for infino & for demonstration/testing purposes.
}
\details{
Description from the author:
Tumor associated macrophages and T cells were characterized in a cohort of 73 
ccRCC and 5 healthy donors using mass cytometry. Macrophage and T cell subsets
of interest were identified and sorted directly from samples known to contain 
a high frequency of these subsets. CD8+ PD-1- T cells (n=6), exhausted CD8+ 
PD-1+ T cells (n=6), control macrophages CD14+ HLA-DR+ CD204- CD38- CD206- (n=11)
and a population of TAM characterized as CD14+ HLA-DR+ CD204+ CD38+ CD206- (n=6) 
were isolated using FACS sorting.
}
\references{
E-MTAB-5640 - RNA sequencing of tumor associated macrophages and T cells in clear cell renal cell carcinoma
(\href{https://www.ebi.ac.uk/arrayexpress/experiments/E-MTAB-5640/}{ArrayExpress})
}
\keyword{datasets}
