% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{run_cibersort}
\alias{run_cibersort}
\title{Run cibersort on expression data}
\usage{
run_cibersort(df, output_file = file.path(tmpdir, "cibersort_output.tsv"),
  error_file = file.path(tmpdir, "cibersort_errors.txt"),
  absolute_mode = FALSE, cibersort_home = "~/cibersort",
  training_data = NULL, genelist = NULL, trans = NULL,
  tmpdir = tempdir(), filter_fun = NULL)
}
\description{
Run cibersort on expression data
}
