% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{run_combat}
\alias{run_combat}
\title{Run combat to normalize for batch effects}
\usage{
run_combat(df, mod = ~1, trans = log1p, batch = "batch")
}
\arguments{
\item{df}{(tbl_df) SxG expression data, with batch stored in sampleinfo attribute}

\item{mod}{(formula) adjustment for any linear covariates in sampleinfo}

\item{trans}{(formula) optional formula to apply to expression data}
}
\description{
Run combat to normalize for batch effects
}
