% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{run_pca}
\alias{run_pca}
\title{run PCA analysis on expression data}
\usage{
run_pca(df, use_ggplot = TRUE, trans = log1p, group = "batch",
  colour = "batch", ...)
}
\arguments{
\item{df}{(tbl_df) GxS expression data, with columns indicating sample_ids}

\item{trans}{(function) if provided, transform expression data prior to PCA}

\item{plot}{(boolean) if TRUE, also plot PCA results}
}
\value{
pca object for each sample run (columns in the provided df)
}
\description{
run PCA analysis on expression data
}
