% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{sampleinfo_to_xdata}
\alias{sampleinfo_to_xdata}
\title{convert celltype column in sampleinfo to xdata design matrix
requires celltype column to have no NAs
run filter_batch first to control which batches will go into this design matrix}
\usage{
sampleinfo_to_xdata(df, sampleinfo = NULL)
}
\arguments{
\item{df}{(tbl_df) SxG df where sampleinfo is optionally stored in sampleinfo attr}

\item{sampleinfo}{optionally pass in sampleinfo here}

\item{batch_names}{(character vector) names of batches to keep}
}
\value{
SxC design matrix
}
\description{
convert celltype column in sampleinfo to xdata design matrix
requires celltype column to have no NAs
run filter_batch first to control which batches will go into this design matrix
}
