% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleorigin_expression-data.R
\docType{data}
\name{singleorigin_expression}
\alias{singleorigin_expression}
\title{Gene expression matrices for immune cells isolated from PBMC samples derived from healthy donors}
\format{an object of class \code{"tbl_df"}}
\usage{
data(singleorigin_expression)
}
\description{
Used as a training dataset for infino & for demonstration/testing purposes.
}
\details{
Description from the author:
(details to come)
}
\keyword{datasets}
