% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{test_cibersort}
\alias{test_cibersort}
\title{Test cibersort using one of the expression matrices provided}
\usage{
test_cibersort(tmpdir = tempdir(), output_file = file.path(tmpdir,
  "cibersort_output.tsv"), test_case = "ExampleMixtures-GEPs.txt",
  cibersort_home = "~/cibersort", input_file = file.path(cibersort_home,
  test_case))
}
\description{
Test cibersort using one of the expression matrices provided
}
