% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{transform_cibersort_to_relative}
\alias{transform_cibersort_to_relative}
\title{Transform 'Absolute' mode of cibersort results to relative proportions}
\usage{
transform_cibersort_to_relative(df)
}
\arguments{
\item{df}{(tbl_df or data.frame) cibersort results formatted by execute_cibersort}
}
\value{
df containing relative proprtions (with absolute score) for each immune cell type
}
\description{
Transform 'Absolute' mode of cibersort results to relative proportions
}
