% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{transform_sampleid}
\alias{transform_sampleid}
\title{Transform sample_id in both expdata & in attributes}
\usage{
transform_sampleid(expdata, fun)
}
\arguments{
\item{expdata}{(tbl_df or data.frame) expression data in GxS format, with sampleinfo attribute}

\item{fun}{(function) function to apply to sample-id column names, and attributes}
}
\value{
transformed expdata object
}
\description{
Transform sample_id in both expdata & in attributes
}
