% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{transpose2gxs}
\alias{transpose2gxs}
\title{Function to transpose an SxG df of expression data into GxS
Optionally applies a numerical transformation (`trans`) to each expression value}
\usage{
transpose2gxs(df, trans = NULL)
}
\arguments{
\item{trans}{(function) optional numerical function to apply to expression data during transform}

\item{SxG}{df (tbl_df or data.frame) containing expression data}
}
\value{
GxS df (tbl_df or data.frame) containing possibly-transformed expression data
}
\description{
Function to transpose an SxG df of expression data into GxS
Optionally applies a numerical transformation (`trans`) to each expression value
}
