% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{transpose_expdata}
\alias{transpose_expdata}
\title{helper function to transpose a dataframe of expression data
The expectation is that that dataframe will have all-columns-but-one be of the same type
and so a matrix-based transpose will be more efficient than one that uses tidyr::gather/spread
(preliminary testing on a single expression matrix indicated a 5x improvement)}
\usage{
transpose_expdata(df, trans = NULL, current_colname, current_rowname)
}
\arguments{
\item{df}{(tbl_df or data.frame) containing expression data to be transposed}

\item{trans}{(function) optional function to apply to expression data during transpose}

\item{current_colname}{(string) name of field to create after transform, holding values of current colnames}

\item{current_rowname}{(string) name of field existing in df, holding current rownames}
}
\value{
transposed df, of type tbl_df
}
\description{
Additionally, while data are in matrix form, one can apply a transformation function
more efficiently than while operating on the full dataframe. Thus, `trans` is supported.
}
