% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{update_sampleinfo}
\alias{update_sampleinfo}
\title{Update sampleinfo affiliated with expdata}
\usage{
update_sampleinfo(expdata, sampleinfo, sample_id = "_SAMPLE_ID")
}
\arguments{
\item{expdata}{(tbl_df or data.frame) sample expression data in GxS format, with attribute sampleinfo}

\item{sampleinfo}{(tbl_df or data.frame) sample-specific info to be merged in with sampleinfo}

\item{sample_id}{(str) name of column in new sampleinfo containing `_SAMPLE_ID`}
}
\description{
Update sampleinfo affiliated with expdata
}
