% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{write_known_samples}
\alias{write_known_samples}
\title{Extract expression data and sampleinfo for all known samples (those where celltype is defined)}
\usage{
write_known_samples(df, expression_fname, xdata_fname)
}
\arguments{
\item{df}{(tbl_df) SxG df where sampleinfo is stored in sampleinfo attr}

\item{expression_fname}{filename to write expression data}

\item{xdata_fname}{filename to write design matrix from sample info}
}
\description{
Extract expression data and sampleinfo for all known samples (those where celltype is defined)
}
