% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_combat.R
\name{write_unknown_samples}
\alias{write_unknown_samples}
\title{Extract expression data and sampleinfo for all unknown samples (those where celltype is undefined)}
\usage{
write_unknown_samples(df, expression_fname)
}
\arguments{
\item{df}{(tbl_df) SxG df where sampleinfo is stored in sampleinfo attr}

\item{expression_fname}{filename to write expression data}
}
\description{
Extract expression data and sampleinfo for all unknown samples (those where celltype is undefined)
}
