% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_estimate.R
\name{run_estimate}
\alias{run_estimate}
\title{Run estimate on expression data}
\usage{
run_estimate(expdf, tmpdir = tempdir(), input_filename = file.path(tmpdir,
  "estimate_input.gct"), output_filename = file.path(tmpdir,
  "estimate_output.gct"))
}
\description{
Run estimate on expression data
}
